ecommerce-django/ecommerce/requirements.txt
Django==4.2.4
django_admin_thumbnails==0.2.8
django_multiupload==0.6.1
reportlab==3.6.8
Requests==2.31.0
xhtml2pdf==0.2.11
